feat: allow overriding symptoms.bsv from outside the container
This adds new support for customizing the dictionary at run time.

You mount a new bsv file into the container at /overrides/symptoms.bsv
like so: type=bind,src=./custom-symptoms/,dst=/overrides/,ro

And upon startup, cTAKES will use a symptoms.bsv dictionary from that
new folder instead of its default Covid one.

Any changes to the file will be noticed and cTAKES will be restarted.
Though note that for this to work correctly, you should bind mount the
directory of the file, not the file itself.

This is mostly for prototyping a new computable phenotype for now,
 but we might extend this further in the future.

The two files that might be interesting to override are:
- symptoms.bsv
- symptoms.xml (which sets up some dictionary parameters)